Use count and size eval together
Evaluate both for each guess, and use one as a tiebreaker for the
other. This put the solver's output more in line with NYT's
evaluation, with count first and size second.

Possible next steps are to allow specifying the ordering of the
evaluation criteria on the command line, and maybe adding the "average
number of possibilities left" metric.