Convert DetailScore to u8 representation
The HashMap in eval_guess was a pretty bad hotspot in profiling. Turns
out you don't need it. There are only 3^5 possible scores = 243, so
it's reasonable to represent scores as a number in [0, 243) and
replace the HashMap with a score-indexed table.

This cut the running time of --solve-all by more than 50%, which I
found shocking -- the HashMap operations did not represent more than
50% of the total running time, and this change actually makes the
biggest hotspot (compute_score) do MORE work, since it now has to
convert the array representation of a score into the number
representation. I only have vague theories for why the difference is
so big. Less allocation? Maybe the array form was being copied around
(seems unlikely)?