Use unchecked array access and wrapping nums in compute_score
compute_score is the hot code, and it was an absolute mess of
branching thanks to array bounds checks. Use get_unchecked for all of
them to avoid this.

The wrapping nums change doesn't matter for release-mode timing since
overflow checks aren't done anyway, but this just makes it clear.