Validate words loaded from file
Since compute_score doesn't do bounds checks anymore, this is required
to make it sound. All arguments ever passed to compute_score are
either references directly into a word list, or to a string that is
equal to a string in the solution list (in the --self-score case).

Also use u8::is_ascii_lowercase elsewhere too.